% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/partR2.R
\name{partR2}
\alias{partR2}
\title{Calculates structure coefficients and commonality coefficients for a gaussian phylogenetic MCMCglmm.}
\usage{
partR2(mod, partvars = NULL, data = NULL, inv_phylo = NULL,
  prior = NULL, nitt = 10000, burnin = 1000, thin = 50)
}
\arguments{
\item{mod}{MCMCglmm model}

\item{partvars}{Predictors for the analysis}

\item{data}{data frame}

\item{inv_phylo}{inverse phylogenetic covariance matrix calcd by MCMCglmm::inverseA}

\item{prior}{prior for MCMCglmm}

\item{nitt}{Number of iterations}

\item{burnin}{Burnin}

\item{thin}{Thinning interval}
}
\value{
Returns a list with two objects
\item{partR2} Mean, Mode and HPD interval of R2
\item{R2_chain} R2 MCMC chain
}
\description{
Calculates structure coefficients and commonality coefficients for a gaussian phylogenetic MCMCglmm.
}
\details{
Not very sophisticated at the moment, just works with this model.
}
\examples{
seals <- data(seal_data)
phylo <- data(seal_phylo)

inv_phylo <- inverseA(seal_phylo, nodes="TIPS",scale=FALSE)$Ainv
prior<-list(G=list(G1=list(V=1,nu=0.002)),R=list(V=1,nu=0.002))

mod_gen <- MCMCglmm(TPM70_ratio ~ num_alleles_mean +  obs_het_mean + prop_low_afs_mean, #
          random=~tip_label, nodes = "TIPS", #   rcov =~us(trait):units
          family=c("gaussian"),ginverse=list(tip_label=inv_phylo),prior=prior,
          data=seal_data,nitt=1100,burnin=100,thin=10)

out <- partR2(mod_gen, partvars = c("num_alleles_mean", "obs_het_mean", "prop_low_afs_mean"),
              data = seal_data, inv_phylo = inv_phylo, prior = prior, nitt = 1100,
              burnin = 100, thin = 10)

}
