% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R2mcmc.R
\name{R2mcmc}
\alias{R2mcmc}
\title{Calculates marginal R2 with CI for MCMCglmm models.}
\usage{
R2mcmc(mod, type = "marginal", family = "gaussian")
}
\arguments{
\item{mod}{MCMCglmm model}

\item{type}{At the moment just "marginal"}

\item{family}{At the moment just "gaussian"}
}
\value{
Returns a list with two objects
\item{partR2} Mean, Mode and HPD interval of R2
\item{R2_chain} R2 MCMC chain
}
\description{
Calculates marginal R2 with CI for MCMCglmm models.
}
\details{
At the moment, the actualy function arguments are defined in the function, except for formula.
}
\examples{
seals <- data(seal_data)
phylo <- data(seal_phylo)

inv_phylo <- inverseA(seal_phylo, nodes="TIPS",scale=FALSE)$Ainv
prior<-list(G=list(G1=list(V=1,nu=0.002)),R=list(V=1,nu=0.002))

mod_gen <- MCMCglmm(TPM70_ratio ~ num_alleles_mean +  obs_het_mean + prop_low_afs_mean,
          random=~tip_label, nodes = "TIPS",
          family=c("gaussian"),ginverse=list(tip_label=inv_phylo),prior=prior,
          data=seal_data,nitt=1100,burnin=100,thin=10)

out <- R2mcmc(mod_gen)
out$partR2


}
